Changed masked filter implementation
The previous implementation based on astropy.convolve and
https://stackoverflow.com/a/36307291, renormalized pixels where the
kernel included masked areas. For example, a uniform_filter with a kernel
equal to [1/3, 1/3, 1/3] applied to a slice [masked, x, y] ended up applying
an effective kernel of [0, 1/2, 1/2]. But this renormalization made the
statistics of those pixels different: some pixels would be the average of
3 pixels, while others of 2 or 1. Hence, they would've had to be thresholded
with a different smo_rv.

The current implementation simply masks all pixels whose calculation involve
a masked value.